pyproject.toml
[project]
name = "aml-v2-deployment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "azure-ai-ml>=1.27.0",
    "azure-identity>=1.22.0",
    "azure-mgmt-resource>=23.3.0",
]
